requirements.txt
Django==5.1.4
djangorestframework==3.15.2
drf-yasg==1.21.8
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
python-decouple==3.8
dj-database-url==2.3.0